# specs/001-title-silosoft-game/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Silosoft Game API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /turn/action/trade:
    post:
      summary: Initiate a trade between two players
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fromPlayer: { type: string }
                toPlayer: { type: string }
                give: { type: array, items: { type: string } }
                receive: { type: array, items: { type: string } }
              required: [fromPlayer, toPlayer]
      responses:
        '200': { description: Trade applied }
  /turn/action/complete:
    post:
      summary: Attempt to complete active feature for a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                playerId: { type: string }
                resourceIds: { type: array, items: { type: string }, minItems: 1 }
                contractorRoles: { type: object, additionalProperties: { type: string } }
              required: [playerId, resourceIds]
      responses:
        '200': { description: Completion result }
...